feat: move account selection codes inside account-selector
