--- ui/pages/settings/transaction-shield-tab/account-selector-modal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './account-selector-modal';
