+++ ui/pages/settings/transaction-shield-tab/account-selector/index.ts
@@ -1,0 +1,1 @@
+export { default } from './account-selector';
